+++ choco/testcentric-gui.nuspec
@@ -60,8 +60,9 @@ Projects with tests to be run under **TestCentric** must already have some versi
     <file src="$BIN$Mono.Options.dll" target="tools" />
     <file src="$BIN$nunit.engine.api.dll" target="tools" />
     <file src="$BIN$testcentric.engine.api.dll" target="tools" />
-    <file src="$BIN$testcentric.engine.metadata.dll" target="tools" />
-    <file src="$BIN$testcentric.engine.core.dll" target="tools" />
+	<file src="$BIN$testcentric.engine.metadata.dll" target="tools" />
+	<file src="$BIN$testcentric.extensibility.dll" target="tools" />
+	<file src="$BIN$testcentric.engine.core.dll" target="tools" />
     <file src="$BIN$testcentric.engine.dll" target="tools" />
 
     <file src="$BIN$Images/RunTests.png" target="tools/Images" />
@@ -91,18 +92,21 @@ Projects with tests to be run under **TestCentric** must already have some versi
     <file src="$BIN$Images/Tree/Visual Studio/Skipped.png" target="tools/Images/Tree/Visual Studio" />
     <file src="$BIN$Images/Tree/Visual Studio/Success.png" target="tools/Images/Tree/Visual Studio" />
 
-    <file src="$BIN$agents/net40/nunit.engine.api.dll" target="tools/agents/net40" />
-    <file src="$BIN$agents/net40/testcentric.engine.metadata.dll" target="tools/agents/net40" />
-    <file src="$BIN$agents/net40/testcentric.engine.core.dll" target="tools/agents/net40" />
-    <file src="$BIN$agents/net40/testcentric-agent.exe" target="tools/agents/net40" />
-    <file src="$BIN$agents/net40/testcentric-agent.exe.config" target="tools/agents/net40" />
-    <file src="$BIN$agents/net40/testcentric-agent-x86.exe" target="tools/agents/net40" />
-    <file src="$BIN$agents/net40/testcentric-agent-x86.exe.config" target="tools/agents/net40" />
-    <file src="testcentric-agent.choco.addins" target="tools/agents/net40" />
+    <file src="$BIN$agents/net462/nunit.engine.api.dll" target="tools/agents/net462" />
+    <file src="$BIN$agents/net462/testcentric.engine.metadata.dll" target="tools/agents/net462" />
+	<file src="$BIN$agents/net462/testcentric.extensibility.dll" target="tools/agents/net462" />
+	<file src="$BIN$agents/net462/testcentric.engine.core.dll" target="tools/agents/net462" />
+	<file src="$BIN$agents/net462/testcentric-agent.exe" target="tools/agents/net462" />
+    <file src="$BIN$agents/net462/testcentric-agent.exe.config" target="tools/agents/net462" />
+    <file src="$BIN$agents/net462/testcentric-agent-x86.exe" target="tools/agents/net462" />
+    <file src="$BIN$agents/net462/testcentric-agent-x86.exe.config" target="tools/agents/net462" />
+    <file src="testcentric-agent.choco.addins" target="tools/agents/net462" />
 
     <file src="$BIN$agents/netcoreapp3.1/nunit.engine.api.dll" target="tools/agents/netcoreapp3.1" />
     <file src="$BIN$agents/netcoreapp3.1/testcentric.engine.metadata.dll" target="tools/agents/netcoreapp3.1" />
-    <file src="$BIN$agents/netcoreapp3.1/testcentric.engine.core.dll" target="tools/agents/netcoreapp3.1" />
+	<file src="$BIN$agents/netcoreapp3.1/testcentric.extensibility.dll" target="tools/agents/netcoreapp3.1" />
+	<file src="$BIN$agents/netcoreapp3.1/Microsoft.Extensions.DependencyModel.dll" target="tools/agents/netcoreapp3.1" />
+	<file src="$BIN$agents/netcoreapp3.1/testcentric.engine.core.dll" target="tools/agents/netcoreapp3.1" />
     <file src="$BIN$agents/netcoreapp3.1/testcentric-agent.dll" target="tools/agents/netcoreapp3.1" />
     <file src="$BIN$agents/netcoreapp3.1/testcentric-agent.dll.config" target="tools/agents/netcoreapp3.1" />
     <file src="$BIN$agents/netcoreapp3.1/testcentric-agent.deps.json" target="tools/agents/netcoreapp3.1" />
@@ -112,7 +116,9 @@ Projects with tests to be run under **TestCentric** must already have some versi
 
     <file src="$BIN$agents/net5.0/nunit.engine.api.dll" target="tools/agents/net5.0" />
     <file src="$BIN$agents/net5.0/testcentric.engine.metadata.dll" target="tools/agents/net5.0" />
-    <file src="$BIN$agents/net5.0/testcentric.engine.core.dll" target="tools/agents/net5.0" />
+	<file src="$BIN$agents/net5.0/testcentric.extensibility.dll" target="tools/agents/net5.0" />
+	<file src="$BIN$agents/net5.0/Microsoft.Extensions.DependencyModel.dll" target="tools/agents/net5.0" />
+	<file src="$BIN$agents/net5.0/testcentric.engine.core.dll" target="tools/agents/net5.0" />
     <file src="$BIN$agents/net5.0/testcentric-agent.dll" target="tools/agents/net5.0" />
     <file src="$BIN$agents/net5.0/testcentric-agent.dll.config" target="tools/agents/net5.0" />
     <file src="$BIN$agents/net5.0/testcentric-agent.deps.json" target="tools/agents/net5.0" />
