+++ nuget/TestCentric.GuiRunner.nuspec
@@ -33,8 +33,9 @@
     <file src="Mono.Options.dll" target="tools" />
     <file src="nunit.engine.api.dll" target="tools" />
     <file src="testcentric.engine.api.dll" target="tools" />
-    <file src="testcentric.engine.metadata.dll" target="tools" />
-    <file src="testcentric.engine.core.dll" target="tools" />
+	<file src="testcentric.engine.metadata.dll" target="tools" />
+	<file src="testcentric.extensibility.dll" target="tools" />
+	<file src="testcentric.engine.core.dll" target="tools" />
     <file src="testcentric.engine.dll" target="tools" />
     <file src="../../nuget/testcentric.nuget.addins" target="tools" />
 
@@ -65,18 +66,21 @@
     <file src="Images/Tree/Visual Studio/Skipped.png" target="tools/Images/Tree/Visual Studio" />
     <file src="Images/Tree/Visual Studio/Success.png" target="tools/Images/Tree/Visual Studio" />
 
-    <file src="agents/net40/nunit.engine.api.dll" target="tools/agents/net40" />
-    <file src="agents/net40/testcentric.engine.metadata.dll" target="tools/agents/net40" />
-    <file src="agents/net40/testcentric.engine.core.dll" target="tools/agents/net40" />
-    <file src="agents/net40/testcentric-agent.exe" target="tools/agents/net40" />
-    <file src="agents/net40/testcentric-agent.exe.config" target="tools/agents/net40" />
-    <file src="agents/net40/testcentric-agent-x86.exe" target="tools/agents/net40" />
-    <file src="agents/net40/testcentric-agent-x86.exe.config" target="tools/agents/net40" />
-    <file src="../../nuget/testcentric-agent.nuget.addins" target="tools/agents/net40" />
+    <file src="agents/net462/nunit.engine.api.dll" target="tools/agents/net462" />
+    <file src="agents/net462/testcentric.engine.metadata.dll" target="tools/agents/net462" />
+	<file src="agents/net462/testcentric.extensibility.dll" target="tools/agents/net462" />
+	<file src="agents/net462/testcentric.engine.core.dll" target="tools/agents/net462" />
+    <file src="agents/net462/testcentric-agent.exe" target="tools/agents/net462" />
+    <file src="agents/net462/testcentric-agent.exe.config" target="tools/agents/net462" />
+    <file src="agents/net462/testcentric-agent-x86.exe" target="tools/agents/net462" />
+    <file src="agents/net462/testcentric-agent-x86.exe.config" target="tools/agents/net462" />
+    <file src="../../nuget/testcentric-agent.nuget.addins" target="tools/agents/net462" />
 
     <file src="agents/netcoreapp3.1/nunit.engine.api.dll" target="tools/agents/netcoreapp3.1" />
     <file src="agents/netcoreapp3.1/testcentric.engine.metadata.dll" target="tools/agents/netcoreapp3.1" />
-    <file src="agents/netcoreapp3.1/testcentric.engine.core.dll" target="tools/agents/netcoreapp3.1" />
+	<file src="agents/netcoreapp3.1/testcentric.extensibility.dll" target="tools/agents/netcoreapp3.1" />
+	<file src="agents/netcoreapp3.1/Microsoft.Extensions.DependencyModel.dll" target="tools/agents/netcoreapp3.1" />
+	<file src="agents/netcoreapp3.1/testcentric.engine.core.dll" target="tools/agents/netcoreapp3.1" />
     <file src="agents/netcoreapp3.1/testcentric-agent.dll" target="tools/agents/netcoreapp3.1" />
     <file src="agents/netcoreapp3.1/testcentric-agent.dll.config" target="tools/agents/netcoreapp3.1" />
     <file src="agents/netcoreapp3.1/testcentric-agent.deps.json" target="tools/agents/netcoreapp3.1" />
@@ -86,7 +90,9 @@
 
     <file src="agents/net5.0/nunit.engine.api.dll" target="tools/agents/net5.0" />
     <file src="agents/net5.0/testcentric.engine.metadata.dll" target="tools/agents/net5.0" />
-    <file src="agents/net5.0/testcentric.engine.core.dll" target="tools/agents/net5.0" />
+	<file src="agents/net5.0/testcentric.extensibility.dll" target="tools/agents/net5.0" />
+	<file src="agents/net5.0/Microsoft.Extensions.DependencyModel.dll" target="tools/agents/net5.0" />
+	<file src="agents/net5.0/testcentric.engine.core.dll" target="tools/agents/net5.0" />
     <file src="agents/net5.0/testcentric-agent.dll" target="tools/agents/net5.0" />
     <file src="agents/net5.0/testcentric-agent.dll.config" target="tools/agents/net5.0" />
     <file src="agents/net5.0/testcentric-agent.deps.json" target="tools/agents/net5.0" />
