+++ src/GuiException/tests/nunit.uiexception.tests.csproj
@@ -2,7 +2,7 @@
   <PropertyGroup>
     <RootNamespace>NUnit.UiException.Tests</RootNamespace>
     <AssemblyName>nunit.uiexception.tests</AssemblyName>
-    <TargetFramework>net45</TargetFramework>
+    <TargetFramework>net462</TargetFramework>
     <OutputPath>..\..\..\bin\$(Configuration)\</OutputPath>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
     <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
@@ -11,7 +11,6 @@
     <Reference Include="System.Windows.Forms" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\tests\test-utilities\test-utilities.csproj" />
     <ProjectReference Include="..\UiException\nunit.uiexception.dll.csproj" />
   </ItemGroup>
   <ItemGroup>
