+++ src/TestCentric/testcentric.exe/TestCentric.Gui.Exe.csproj
@@ -3,7 +3,7 @@
 		<OutputType>WinExe</OutputType>
 		<RootNamespace>TestCentric.Gui</RootNamespace>
 		<AssemblyName>testcentric</AssemblyName>
-		<TargetFramework>net45</TargetFramework>
+		<TargetFramework>net462</TargetFramework>
 		<OutputPath>..\..\..\bin\$(Configuration)\</OutputPath>
 		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
 		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
