+++ src/TestCentric/tests/Presenters/Main/WhenTestRunCompletes.cs
@@ -12,8 +12,6 @@ namespace TestCentric.Gui.Presenters.Main
 
     public class WhenTestRunCompletes : MainPresenterTestBase
     {
-        private ResultSummary _summary;
-
         [SetUp]
         public void SimulateTestRunFinish()
         {
@@ -23,7 +21,7 @@ namespace TestCentric.Gui.Presenters.Main
             _model.HasResults.Returns(true);
             _model.ResultSummary.Returns(new ResultSummary() { FailureCount = 1 });
             _model.IsTestRunning.Returns(false);
-            _model.TestPackage.Returns(new NUnit.Engine.TestPackage("dummy.dll"));
+            _model.TestPackage.Returns(new NUnit.Engine.TestPackage(new[] { "dummy.dll" }));
 
             var resultNode = new ResultNode("<test-run id='XXX' result='Failed' />");
             FireRunFinishedEvent(resultNode);
