+++ src/TestCentric/tests/Presenters/Main/WhenTestsAreReloaded.cs
@@ -22,7 +22,7 @@ namespace TestCentric.Gui.Presenters.Main
 
             TestNode testNode = new TestNode("<test-suite id='1'/>");
             _model.LoadedTests.Returns(testNode);
-            _model.TestPackage.Returns(new NUnit.Engine.TestPackage("dummy.dll"));
+            _model.TestPackage.Returns(new NUnit.Engine.TestPackage(new[] { "dummy.dll" }));
 
             FireTestReloadedEvent(testNode);
         }
