+++ src/TestCentric/tests/Presenters/NUnitTreeDisplayStrategyTests.cs
@@ -25,7 +25,7 @@ namespace TestCentric.Gui.Presenters.TestTree
         {
             _view = Substitute.For<ITestTreeView>();
             _model = Substitute.For<ITestModel>();
-            _settings = new TestCentric.TestUtilities.Fakes.UserSettings();
+            _settings = new Fakes.UserSettings();
             _model.Settings.Returns(_settings);
             _model.TestFiles.Returns(new List<string>());
