+++ src/TestCentric/tests/Views/ErrorsAndFailuresViewTests.cs
@@ -5,7 +5,6 @@
 
 using System.Windows.Forms;
 using NUnit.Framework;
-using TestCentric.TestUtilities;
 
 namespace TestCentric.Gui.Views
 {
