+++ src/TestModel/model/TestCentric.Gui.Model.csproj
@@ -3,7 +3,7 @@
 	<PropertyGroup>
 		<RootNamespace>TestCentric.Gui.Model</RootNamespace>
 		<AssemblyName>TestCentric.Gui.Model</AssemblyName>
-		<TargetFramework>net45</TargetFramework>
+		<TargetFramework>net462</TargetFramework>
 		<OutputPath>..\..\..\bin\$(Configuration)\</OutputPath>
 		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
 		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
@@ -14,10 +14,10 @@
 	</ItemGroup>
   
 	<ItemGroup>
-    <PackageReference Include="Mono.Options" Version="6.12.0.148" />
-    <PackageReference Include="NUnit.Engine.Api" Version="4.0.0-dev00102" />
-    <PackageReference Include="TestCentric.Engine.Api" Version="2.0.0-dev00002" />
-    <PackageReference Include="TestCentric.Engine" Version="2.0.0-dev00002" />
-  </ItemGroup>
+		<PackageReference Include="Mono.Options" Version="6.12.0.148" />
+		<PackageReference Include="NUnit.Engine.Api" Version="3.16.2" />
+		<PackageReference Include="TestCentric.Engine.Api" Version="2.0.0-dev00023" />
+		<PackageReference Include="TestCentric.Engine" Version="2.0.0-dev00023" />
+	</ItemGroup>
   
 </Project>
