+++ src/TestModel/tests/Fakes/ServiceLocator.cs
@@ -7,7 +7,7 @@ using System;
 using System.Collections.Generic;
 using NUnit.Engine;
 
-namespace TestCentric.TestUtilities.Fakes
+namespace TestCentric.Gui.Model.Fakes
 {
     public class ServiceLocator : IServiceLocator
     {
