+++ src/TestModel/tests/MakeTestPackageTests.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
 using TestCentric.Common;
-using TestCentric.TestUtilities.Fakes;
+using TestCentric.Gui.Model.Fakes;
 
 using TestPackage = NUnit.Engine.TestPackage;
