+++ src/TestModel/tests/TestCentric.Gui.Model.Tests.csproj
@@ -2,7 +2,7 @@
 	<PropertyGroup>
 		<RootNamespace>TestCentric.Gui.Model</RootNamespace>
 		<AssemblyName>TestCentric.Gui.Model.Tests</AssemblyName>
-		<TargetFramework>net45</TargetFramework>
+		<TargetFramework>net462</TargetFramework>
 		<OutputPath>..\..\..\bin\$(Configuration)\</OutputPath>
 		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
 		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
@@ -14,9 +14,6 @@
         <ProjectReference Include="..\..\tests\mock-assembly\mock-assembly.csproj">
           <Name>mock-assembly</Name>
         </ProjectReference>
-        <ProjectReference Include="..\..\tests\test-utilities\test-utilities.csproj">
-          <Name>test-utilities</Name>
-        </ProjectReference>
         <ProjectReference Include="..\model\TestCentric.Gui.Model.csproj">
           <Name>TestCentric.Gui.Model</Name>
         </ProjectReference>
@@ -24,7 +21,7 @@
     <ItemGroup>
         <PackageReference Include="NSubstitute" Version="4.0.0" />
         <PackageReference Include="NUnit" Version="3.13.2" />
-        <PackageReference Include="NUnit.Engine.Api" Version="4.0.0-dev00102" />
-        <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.1" />
-    </ItemGroup>
+        <PackageReference Include="NUnit.Engine.Api" Version="3.16.2" />
+		<PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.1" />
+	</ItemGroup>
 </Project>
