--- src/tests/test-utilities/Properties/AssemblyInfo.cs
@@ -1,6 +1,0 @@
-// ***********************************************************************
-// Copyright (c) Charlie Poole and TestCentric GUI contributors.
-// Licensed under the MIT License. See LICENSE file in root directory.
-// ***********************************************************************
-
-[assembly: NUnit.Framework.NonTestAssembly]
